+++ server/models/campaign.py
@@ -30,6 +30,7 @@ class Room(BaseModel):
     invitation_code = TextField(default=uuid.uuid4, unique=True)
     player_location = TextField(null=True)
     dm_location = TextField(null=True)
+    is_locked = BooleanField(default=False)
 
     def __repr__(self):
         return f"<Room {self.get_path()}>"